Using circular queue
