+++ CHANGELOG.md
@@ -7,6 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## Changes
+
+- Swapping the lock-free Queue out with a simpler locking queue that has significantly less lock contention in scenarios
+  when multiple buffers are required.
+
 ## [v0.3.1] - 2022-03-12 (Beta)
 
 ### Fixes
@@ -172,27 +177,16 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Initial Release of Frisbee
 
 [unreleased]: https://github.com/loopholelabs/frisbee/compare/v0.3.1...HEAD
-
 [v0.3.1]: https://github.com/loopholelabs/frisbee/compare/v0.3.0...v0.3.1
-
 [v0.3.0]: https://github.com/loopholelabs/frisbee/compare/v0.2.4...v0.3.0
-
 [v0.2.4]: https://github.com/loopholelabs/frisbee/compare/v0.2.3...v0.2.4
-
 [v0.2.3]: https://github.com/loopholelabs/frisbee/compare/v0.2.2...v0.2.3
-
 [v0.2.2]: https://github.com/loopholelabs/frisbee/compare/v0.2.1...v0.2.2
-
 [v0.2.1]: https://github.com/loopholelabs/frisbee/compare/v0.2.0...v0.2.1
-
 [v0.2.0]: https://github.com/loopholelabs/frisbee/compare/v0.1.6...v0.2.0
-
 [v0.1.6]: https://github.com/loopholelabs/frisbee/compare/v0.1.5...v0.1.6
-
 [v0.1.5]: https://github.com/loopholelabs/frisbee/compare/v0.1.4...v0.1.5
-
 [v0.1.4]: https://github.com/loopholelabs/frisbee/compare/v0.1.3...v0.1.4
-
 [v0.1.3]: https://github.com/loopholelabs/frisbee/compare/v0.1.2...v0.1.3
 [v0.1.2]: https://github.com/loopholelabs/frisbee/compare/v0.1.1...v0.1.2
 [v0.1.1]: https://github.com/loopholelabs/frisbee/compare/v0.1.0...v0.1.1
